I2CDevObj: just use a normal unsigned
No need to optimize these things, better to use defaults and let the
compiler optimize.

+++ framework/include/I2CDevObj.hpp
@@ -83,7 +83,7 @@ protected:
 			    uint32_t transfer_timeout_usec);
 
 	int m_fd;
-	uint16_t _retries;
+	unsigned _retries;
 };
 
 };
